refactor: modularize response generation by creating buildMessages, executeFunctionCalls, and StreamProcessor classes
